requirements.txt
absl-py==2.1.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cssselect==1.2.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.4
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
fitz==0.0.1.dev2
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googlesearch-python==1.2.5
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
libclang==18.1.1
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.13.3
networkx==3.4.2
newspaper3k==0.2.8
nibabel==5.3.2
nipype==1.9.0
nltk==3.9.1
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
protobuf==5.28.3
prov==2.0.1
puremagic==1.28
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scipy==1.14.1
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
streamlit==1.40.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tesseract==0.1.3
tf_keras==2.18.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traits==6.4.3
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
wrapt==1.16.0
